Make featurizers and unit tests compile and run with GTest.   Create definitions for the first featurizer kernel.   Add new operator domain.   Create datetime_transformer kernel and build.   Move OPAQUE types definitions for featurizers kerneles out to a separate cc.   Register them with the type system.

+++ onnxruntime/core/automl/featurizers/src/FeaturizerPrep/Featurizers/SampleAdd.cpp
@@ -32,7 +32,7 @@ Estimator & Estimator::fit_impl(apache_arrow const &data) /*override*/ {
 }
 
 Estimator::TransformerUniquePtr Estimator::commit_impl(void) /*override*/ {
-    return std::make_unique<SampleAdd::Transformer>(_accumulated_delta);
+  return std::make_unique<SampleAdd::Transformer>(static_cast<std::uint16_t>(_accumulated_delta));
 }
 
 } // namespace SampleAdd
